Game now seems to run properly
The bug was caused by the testMove() Method, it did copy the Game, but the copy still had references to the original map,
adding additional figures and allowing illegal moves.

Instead of the copyGame method, most classes now implement the Cloneable interface,
although not in the proper way, since they construct new Objects instead of calling super.clone().

TODO: Castling is not implemented yet and a few other rules are missing as well